Code refactoring 3

+++ routes/web/webAuthors.php
@@ -11,8 +11,8 @@ Route::middleware(['auth'])->prefix('authors')->name('authors.')->group(function
     Route::get('/', GetListAuthorWebController::class)->name('index');
     Route::get('/create', [CreateAuthorWebController::class, 'create'])->name('create');
     Route::post('/', [CreateAuthorWebController::class, 'store'])->name('store');
-    Route::get('/{author}', GetByIdAuthorWebController::class)->name('show');
-    Route::get('/{author}/edit', [UpdateAuthorWebController::class, 'edit'])->name('edit');
-    Route::put('/{author}', [UpdateAuthorWebController::class, 'update'])->name('update');
-    Route::delete('/{author}', DeleteAuthorWebController::class)->name('destroy');
+    Route::get('/{author}', GetByIdAuthorWebController::class)->name('show')->whereNumber('author');
+    Route::get('/{author}/edit', [UpdateAuthorWebController::class, 'edit'])->name('edit')->whereNumber('author');
+    Route::put('/{author}', [UpdateAuthorWebController::class, 'update'])->name('update')->whereNumber('author');
+    Route::delete('/{author}', DeleteAuthorWebController::class)->name('destroy')->whereNumber('author');
 });
